gitk: Visualize a merge commit w/ right-click in gitk
When first doing a merge in git-gui, the "Visualize Merge" button is
quite helpful.  But once the merge is complete, its quite hard, especially
for a novice, to recreate that merge view.

This commit adds to gitk the ability to right-click on a merge and bring
up a new gitk window displaying only those commits touched by a merge.

Signed-off-by: Jason Holden <jason.k.holden@gmail.com>